move KDF interface
